Move certificate mock to another package
